--- Form2.cs
@@ -1,32 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace WindowsFormsApp1
-{
-    public partial class Form2 : Form
-    {
-        public Form2()
-        {
-            InitializeComponent();
-        }
-
-        private void button1_Click(object sender, EventArgs e)
-        {
-            Form1 f = new Form1();
-            f.Show();   
-            this.button1.Enabled = false;
-            this.Visible = false;
-        }
-        public void ShutDown()
-        {
-            this.Dispose();
-        }
-    }
-}
